feature/project-data -> Manage team members in a project

+++ src/main/java/com/erd/core/service/ProjectService.java
@@ -3,6 +3,7 @@ package com.erd.core.service;
 import com.erd.core.dto.request.ProjectCreateRequestDTO;
 import com.erd.core.dto.request.ProjectUpdateRequestDTO;
 import com.erd.core.dto.request.TeamMemberRequestDTO;
+import com.erd.core.dto.request.UpdateTeamMemberRequestDTO;
 import com.erd.core.dto.response.ProjectDetailsResponseDTO;
 import com.erd.core.dto.response.ProjectResponseDTO;
 import com.erd.core.model.Project;
@@ -25,11 +26,13 @@ public class ProjectService {
 
     private final ProjectRepository projectRepository;
     private final TeamService teamService;
+    private final UserService userService;
     private final ModelMapper modelMapper;
 
-    public ProjectService(ProjectRepository projectRepository, TeamService teamService, ModelMapper modelMapper) {
+    public ProjectService(ProjectRepository projectRepository, TeamService teamService, UserService userService, ModelMapper modelMapper) {
         this.projectRepository = projectRepository;
         this.teamService = teamService;
+        this.userService = userService;
         this.modelMapper = modelMapper;
     }
 
@@ -65,6 +68,11 @@ public class ProjectService {
             throw new RuntimeException("Project not found for id: " + projectUpdateRequestDto.getId());
         }
 
+        var userId = userService.getUserIdByLoggedUserEmail();
+        if (teamService.isUserOwner(userId, projectUpdateRequestDto.getId())) {
+            throw new RuntimeException("Only the OWNER can update the project");
+        }
+
         logger.info("Updating project data");
         Project updatedProject = projectRepository.save(modelMapper.map(projectUpdateRequestDto, Project.class));
         return modelMapper.map(updatedProject, ProjectResponseDTO.class);
@@ -73,13 +81,43 @@ public class ProjectService {
     public void addTeamMember(TeamMemberRequestDTO teamMemberRequestDto) {
         UUID projectId = teamMemberRequestDto.getProjectId();
         logger.info("Finding project by id: {}", projectId);
+
+        var userId = userService.getUserIdByLoggedUserEmail();
+        if (teamService.isUserOwner(userId, projectId)) {
+            throw new RuntimeException("Only the OWNER can add team members to the project");
+        }
+
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(() -> new RuntimeException("Project not found for id: " + projectId));
         teamService.addTeamMember(teamMemberRequestDto, project);
     }
 
+    public void updateTeamMember(UpdateTeamMemberRequestDTO requestDTO) {
+        var userId = userService.getUserIdByLoggedUserEmail();
+        if (teamService.isUserOwner(userId, requestDTO.getProjectId())) {
+            throw new RuntimeException("Only the OWNER can update team members");
+        }
+
+        teamService.updateTeamMember(requestDTO);
+    }
+
+    public void removeTeamMember(UUID memberId, UUID projectId) {
+        var userId = userService.getUserIdByLoggedUserEmail();
+        if (teamService.isUserOwner(userId, projectId)) {
+            throw new RuntimeException("Only the OWNER can remove team members");
+        }
+
+        teamService.removeTeamMember(memberId, projectId);
+    }
+
     public void deleteById(UUID id) {
         logger.info("Deleting project by id: {}", id);
+
+        var userId = userService.getUserIdByLoggedUserEmail();
+        if (teamService.isUserOwner(userId, id)) {
+            throw new RuntimeException("Only the OWNER can delete the project");
+        }
+
         projectRepository.deleteById(id);
         // TODO: delete diagram data
     }
